refactor(js-sdk): make utils.wait abort-aware and reuse it in retry
